Aslan Lexer: Just the comments
I am commiting the structure of files based on the URI:
http://www.phpcompiler.org/articles/reentrantparser.html. I need
think a good way of divide it and put some sort of testing.

One way of test the language is to make a input file and check it
by a grammar's test. If I do this to the lexer part, how I will test
the real grammar? A unique way of testing is best.