refactory
I'm changing to use CMake and GNU Make. So, I removed the GNU makefile
and now I am generating it. Also, I moved the files to others directory
and decided to make a directory called tests (for now, empty). Other
decision was not use gtest/gmock because it conflicts with CMake test
option.