putting the code on drone.io
To put the code on drone.io was necessary to taking out -Werror and
other options from my code. I dont know if it is a problem from bison
or gcc newer, but I will pass to show the warnings from now on. I hope
this wont blind me. In test files, others minor changes. A missing
include added (cstring for strcmp) and I changed the fprintf to cerr.
I could included the stdio.h, but it will be less C++.