Initial commit of a grammar
This first commit create the first part of the AgentSpeak Language.
It is just the recognize of the text without any action and I've to
stop the developing in the expression non-terminals. The structure
to recognize stuff like: "beliefThatIsTrue.", "~beliefThatIsFalse.",
"!do.", "+!do.", "-?beliefThatIsFalse." and "+belief.".

What this stuff does? Well, it does nothing. :)

A better answer is that in the future, it will be translated to some
other language. This language will have a framework that will accept
the code generated to operate. The meaning of the stuff is describe
in at least two works: [1] and [2].

[1] AgentSpeak (L): BDI agents speak out in a logical computable language
[2] Programming multi-agent systems in AgentSpeak using Jason