Crafting a symbols' table
I decided make a table of symbols because I need to retrieve and store
the symbols of grammar that I have found. Although I do not believe it
is enough, it is much better than my first version using a tree.
The gist of this is based in an article in game programming gems, but I
don't remember to put a link here -.-'