Changing symbol field in union and fixing literals
The literal rule for some reason was returning a symbol type, this
rule need to return a lexema type (a C string). The union is been
changed in symbol which is a void pointer to a Symbol pointer. To
do this I need include AllSymbol.hpp in grammar file, lexer file
and general test header file. Also the mock classes had the base
class changed to Symbol class.