Type value of Symbol comes from a enum
Changing to make the type value of Symbol class that was hard-coded
to comes from a enum. I could have changed the type of the type field
to depend on this new enum, but I prefer not do it because it will
prevent some extensions from the user side. What I mean by that? Lets
say that I'd like to make a T class and extend Symbol too. I can
put a value type that the system is not using right now. If I change
the type, the user will need change the enum. So, it is just to prevent
people that need to do patches to have to do it again because some
weird conflict...